Optimize stack maps: add fast path for no inline info.
Consumers of CodeInfo can skip significant chunks of work
if they can quickly determine that method has no inlining.

Store this fact as a flag bit at the start of code info.
This changes binary format and adds <0.1% to oat size.

I added the extra flag field as the simplest solution for now,
although I would like to use it for more things in the future.
(e.g. store the special cases of empty/deduped tables in it)

This improves app startup by 0.4% (maps,speed).
PMD on golem seems to gets around 15% faster.

Bug: 133257467
Test: ./art/test.py -b --host --64
Change-Id: Ia498a31bafc74b51cc95b8c70cf1da4b0e3d894e